Extract transcript-raw module and write CLI transcript at agent runner level
