+++ src/config/sessions.ts
@@ -7,3 +7,4 @@ export * from "./sessions/session-key.js";
 export * from "./sessions/store.js";
 export * from "./sessions/types.js";
 export * from "./sessions/transcript.js";
+export * from "./sessions/transcript-raw.js";
